Agregué vista venta e hice varios cambios a los metodos

+++ src/inventario/AdministradorCatalogo.java
@@ -39,7 +39,7 @@ public class AdministradorCatalogo {
      
     }
     
-    public void modificarProducto(int cod, String descripcion, int precio, int stock, int stockMin){
+    public void modificarProducto(int cod, String descripcion, float precio, int stock, int stockMin){
         
         Producto p = buscarProducto(cod);
